+++ src/Helper/SerializationHelper.php
@@ -51,9 +51,13 @@ class SerializationHelper extends AbstractHelper
     public static function serialize($object)
     {
         $items = array();
-        if ($object instanceof Traversable) {
+        if ($object instanceof Traversable || is_array($object)) {
             foreach ($object as $item) {
-                $items[] = self::toArray($item);
+                if (is_object($item)) {
+                    $items[] = self::toArray($item);
+                } else {
+                    $items[] = $item;
+                }
             }
             return json_encode($items);
         } else {